Modifiche del Codice. Test e  Interfacce

+++ Codice/23.06.2021 IDS/app/src/main/java/it/unicam/cs/ids2021/negozio/Prodotto.java
@@ -1,5 +1,5 @@
 package it.unicam.cs.ids2021.negozio;
 
 public interface Prodotto {
-    float getVolume();
+    double getVolume();
 }
